Add domain items

+++ internal/repository/repository.go
@@ -7,8 +7,8 @@ import (
 
 type UserRepository interface {
 	List(ctx context.Context, paging model.Paging) ([]model.User, int64, error)
-	GetById(ctx context.Context, id uint) (*model.User, error)
-	Create(ctx context.Context, user *model.User) (uint, error)
+	GetById(ctx context.Context, id uint64) (*model.User, error)
+	Create(ctx context.Context, user *model.User) (uint64, error)
 	Update(ctx context.Context, user *model.User) error
-	ForceDelete(ctx context.Context, id uint) error
+	ForceDelete(ctx context.Context, id uint64) error
 }
